cli/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.4
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
requests==2.32.4
rich==14.1.0
shellingham==1.5.4
typer==0.16.0
typing_extensions==4.14.1
urllib3==2.5.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pycparser==2.22
Pygments==2.19.2
PyNaCl==1.5.0
requests==2.32.4
rich==14.1.0
shellingham==1.5.4
typer==0.16.0
typing_extensions==4.14.1
urllib3==2.5.0
